Write getters for struct Game and tests for game_add method
In order to write tests for game_add, we needed an access to internal values in Game. We wrote getters for these values, and a set of tests for the game_add function. The player names are currently stocked in a bugged array, we will change that when we will rethink the architecture of the Game class.